delete guifont
Because it can not ensure every one has this font. In fact, if you dont has this font, it will result an error. Even if you wanna use a kind of font which is as same as visual studio, you should use "set guifont=Consolas_NF:h10:cANSI:qDRAFT". Because nerd-fonts have substitute powerfonts.